% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeres.R
\name{change.res}
\alias{change.res}
\title{Change resolution of a raster}
\usage{
change.res(
  x,
  targ.res = NULL,
  method = c("mean", "max", "min", "median", "sum", "modal", "near", "bilinear")
)
}
\arguments{
\item{x}{A terra SpatRaster or raster RasterLayer object}

\item{targ.res}{Positive integer. Desired resolution of output raster. Single
value asserts equal horz/vert length. Or two integers for
different horz/vert lengths.}

\item{method}{method for dis-/aggregating. one of c('mean', 'max', 'min',
'median', 'sum', 'modal') for aggregation or one of c('near',
'bilinear') for disaggregation}
}
\value{
A raster of the resolution stated in target.res
}
\description{
Changes the resolution of a raster through dis-/aggregation
}
\examples{
\dontrun{
x <- rast()
values(x) <- 1:ncell(x)
y <- raster(x)
x <- changeres(x, 20, method='median')
y <- changeres(y, 20, method='median')
x <- changeres(x, 1, method='bilinear')
y <- changeres(y, 1, method='bilinear')
}
}
