% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cos.aspect.R
\name{cosaspect}
\alias{cosaspect}
\title{Cosine-transform aspect}
\usage{
cosaspect(x, unit = c("rad", "deg"))
}
\arguments{
\item{x}{A vector, dataframe, or raster of aspect values}

\item{unit}{Character. Units that aspect are in natively, either degrees or radians}
}
\value{
A vector, dataframe, or raster of cosine-transformed aspect values
}
\description{
Transforms simple cardinal aspect values into their cosines
}
\examples{
\dontrun{
library(terra)
crb <- readRDS(system.file(file.path('data', 'crb.rds'), package='topographer'))
a <- terra::terrain(crb, v='aspect')
u <- 'deg'
c.aspect <- cosaspect(a, u)
plot(c.aspect)
}
}
