% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.rad.convert.R
\name{d2r}
\alias{d2r}
\title{Convert degrees to radians or radians to degrees}
\usage{
d2r(a)
}
\arguments{
\item{a}{a float angle in degrees or radians}
}
\value{
the angle converted into degrees or radians
}
\description{
calculates zonal statistics for raster inputs given an input shapefile or coordinates
}
\examples{
\dontrun{
d <- 25.5
r <- d2r(d)
d <- r2d(r)
}
}
