% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspect.transform.R
\name{facing}
\alias{facing}
\title{Transform aspect}
\usage{
facing(slope, aspect, focal = 180, unit = "rad")
}
\arguments{
\item{slope}{vector, dataframe or raster of slope values}

\item{aspect}{vector, dataframe, or raster of aspect values}

\item{focal}{focal angle for facingness}

\item{unit}{units that aspect are in, either degrees or radians}

\item{fold}{angle along which aspect should be folded}
}
\value{
transformed aspect values
}
\description{
transforms simple cardinal aspect values into ecologically meaningful values
}
\examples{
\dontrun{
d <- 25.5
r <- d2r(d)
d <- r2d(r)
}
}
