% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold.aspect.R
\name{fold.aspect}
\alias{fold.aspect}
\title{Fold aspect about a given focal angle}
\usage{
fold.aspect(x, fold = 180)
}
\arguments{
\item{x}{A vector, dataframe, or raster of aspect values}

\item{fold}{Numeric. The angle along which aspect should be folded, default=180}
}
\value{
A vector, dataframe, or raster of folded aspect values
}
\description{
Transforms simple cardinal aspect values into values folded along a selected axis
}
\examples{
\dontrun{
library(terra)
crb <- readRDS(system.file(file.path('data', 'crb.rds'), package='topographer'))
a <- terra::terrain(crb, v='aspect')
f <- 225
f.aspect <- fold.aspect(a, f)
plot(f.aspect)
}
}
