% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.rasters.R
\name{get.rasters}
\alias{get.rasters}
\title{Source rasters from local data directory}
\usage{
get.rasters(x, dir)
}
\arguments{
\item{x}{a string containing a search pattern for a subdirectory name indicating the
raster data of interest}

\item{dir}{a string containing a directory path where rasters are stored}
}
\value{
a list of SpatRaster objects for serial processing
}
\description{
Sources rasters from a local directory using terra::rast and stores in a list of SpatRaster objects.
}
\examples{
\dontrun{
rasdir <- './data'
ras.ls <- get.rasters('raster', rasdir)
}
}
