% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.rad.convert.R
\name{r2d}
\alias{r2d}
\title{Convert radians to degrees}
\usage{
r2d(a)
}
\arguments{
\item{a}{a float angle in degrees or radians}
}
\value{
the angle converted into degrees or radians
}
\description{
calculates zonal statistics for raster inputs given an input shapefile or coordinates
}
\examples{
\dontrun{
d <- 25.5
r <- d2r(d)
d <- r2d(r)
}
}
