% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatload.R
\name{thl}
\alias{thl}
\title{Heat load}
\usage{
thl(L, A, S, unit = "deg", fold = 180)
}
\arguments{
\item{L}{float value for latitude of site}

\item{A}{vector, dataframe, or raster of aspect values}

\item{S}{vector, dataframe or raster of slope values}

\item{unit}{units that aspect are in, either degrees or radians}

\item{fold}{angle along which aspect should be folded}
}
\value{
total heat load value
}
\description{
calculates heat load from input slope and aspect values
}
\note{
Heat load is a unitless index of heat load, rather than a direct estimate of
incident radiation. The approach implements Equation 3 from McCune and Keon (2002),
using inputs of latitude. The equation holds between latitudes of ±30-60º and
slope angles <= 60º. Their 3rd equation is valid for the narrowest input ranges
but has the highest R^2 and is suitable for many settings. I may implement
equations 1 and 2 in the future. NOTE: the estimates are somewhat rudimentary
and do not take into account cloud cover, regional differences in the atmospheric
coefficient, or shading by adjacent topography.
}
\examples{
\dontrun{
l <- 38.58
a <- 0:360
s <- 12.6
heatload <- thl(l, a, s, 'deg', 180)
plot(A,thl(L,A,S,fold=225))
points(A,thl(L,A,S,fold=225))
}
}
\references{
McCune, B. and D. Keon (2002) Equations for potential annual direct incident
radiation and heat load. Journal of Vegetation Science 13:603-606.
}
