% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twi.R, R/uca.R
\name{twi}
\alias{twi}
\title{Topographic Wetness Index (TWI)}
\usage{
twi(x, in.res = 100, out.res = 10, fill.sinks = T, deg = 0.1)

twi(x, in.res = 100, out.res = 10, fill.sinks = T, deg = 0.1)
}
\arguments{
\item{x}{A terra SpatRaster or raster RasterLayer object}

\item{in.res}{Positive integer. Resolution for computing upslope area
and localslope.  Single value asserts equal horz/vert length.
Or two integers for different horz/vert lengths.}

\item{out.res}{Positive integer. Desired resolution of output raster. Single
value asserts equal horz/vert length. Or two integers for
different horz/vert lengths.}

\item{deg}{Threshold intercell slope to determine sinks}

\item{fill_sinks}{Logical. If TRUE (default) then run a sinkfill before
calculating the upslope area and TWI.}
}
\value{
A raster comprising topographic wetness index values

A raster comprising topographic wetness index values
}
\description{
calculates Topographic Wetness Index on a raster DEM

calculates Upslope Contributing Area on a raster DEM
}
\examples{
\dontrun{
x <- rast()
values(x) <- 1:ncell(x)
y <- raster(x)
x.twi <- twi(x, in.res=100, out.res=10, fill.sinks=T)
y.twi <- twi(y, in.res=100, out.res=1000, fill.sinks=T)
plot(x.twi)
}
\dontrun{
x <- rast()
values(x) <- 1:ncell(x)
y <- raster(x)
x.twi <- twi(x, in.res=100, out.res=10, fill.sinks=T)
y.twi <- twi(y, in.res=100, out.res=1000, fill.sinks=T)
plot(x.twi)
}
}
