% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uca.R
\name{uca}
\alias{uca}
\title{Upslope contributing area}
\usage{
uca(x, calc.res = 100, out.res = 10, fill.sinks = T, deg = 0.1)
}
\arguments{
\item{x}{A terra SpatRaster or raster RasterLayer object}

\item{calc.res}{Positive integer. Resolution for computing upslope area
and localslope.  Single value asserts equal horz/vert length.
Or two integers for different horz/vert lengths.}

\item{out.res}{Positive integer. Desired resolution of output raster. Single
value asserts equal horz/vert length. Or two integers for
different horz/vert lengths.}

\item{deg}{Threshold intercell slope to determine sinks}

\item{fill_sinks}{Logical. If TRUE (default) then run a sinkfill before
calculating the upslope area and TWI.}
}
\value{
A raster comprising topographic wetness index values
}
\description{
calculates upslope contributing area per pixel on a raster DEM
}
\examples{
\dontrun{
library(topographer)
crb <- readRDS(file.path('topographer', 'data', 'crb.rds'))
crb <- terra::unwrap(crb)
crb <- raster(crb)
y <- raster(crb)
values(y) <- values(crb)
x.uca <- uca(crb, calc.res=100, out.res=10, fill.sinks=T)
y.uca <- uca(y, calc.res=100, out.res=1000, fill.sinks=T)
plot(x.uca)
plot(y.uca)
}
}
