% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspect.transform.R
\name{facing}
\alias{facing}
\title{Transform aspect}
\usage{
facing(slope, aspect, focal = 180, unit = "rad")
}
\arguments{
\item{slope}{A vector, dataframe or raster of slope values}

\item{aspect}{A vector, dataframe, or raster of aspect values}

\item{focal}{Integer. A focal angle for transforming aspect into a 'facing' value}

\item{unit}{Character. Units that aspect are in natively, either degrees or radians}

\item{fold}{angle along which aspect should be folded}
}
\value{
transformed aspect values
}
\description{
Transforms simple cardinal aspect values into ecologically meaningful values
}
\examples{
\dontrun{
crb <- readRDS(file.path('data', 'crb.rds'))
s <- terra::terrain(crb, v='slope')
a <- terra::terrain(crb, v='aspect')
f <- 180
u <- 'deg'
e.facing <- eastness(s,a,u)
adjs.facing <- adjsouthness(slope=s,aspect=a,focal=f,unit=u)
f.aspect <- fold.aspect(a, fold=215)
s.aspect <- sin.aspect(a)
c.aspect <- cos.aspect(a)
plot(a)
plot(e.facing)
plot(f.aspect)
}
}
