% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facing.R
\name{facing}
\alias{facing}
\title{Facing aspect}
\usage{
facing(slope, aspect, focal = 180, unit = "rad")
}
\arguments{
\item{slope}{A vector, dataframe or raster of slope values}

\item{aspect}{A vector, dataframe, or raster of aspect values}

\item{focal}{Integer. A focal angle for transforming aspect into a 'facing' value}

\item{unit}{Character. Units that aspect are in natively, either degrees or radians}

\item{fold}{Numeric. Angle along which aspect should be folded}
}
\value{
A vector, dataframe, or raster of facing-transformed aspect values
}
\description{
Transforms simple cardinal aspect values into values relative to a focal direction
}
\examples{
\dontrun{
library(terra)
crb <- readRDS(system.file(file.path('data', 'crb.rds'), package='topographer'))
s <- terra::terrain(crb, v='slope')
a <- terra::terrain(crb, v='aspect')
f <- 180
u <- 'deg'
facing.270 <- facing(s,a,270,u)
plot(facing.270)
}
}
