% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.rasters.R
\name{get.rasters}
\alias{get.rasters}
\title{Ingest rasters from data directory}
\usage{
get.rasters(x, path)
}
\arguments{
\item{x}{Character vector. Search pattern for the raster data of interest}

\item{path}{Character vector. Directory path where rasters are stored}
}
\value{
a list of SpatRaster objects for serial processing
}
\description{
Sources rasters from a local directory using terra::rast and
stores in a list of SpatRaster objects.
}
\examples{
\dontrun{
rasdir <- './data'
ras.ls <- get.rasters('raster', rasdir)
}
}
