% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatload.R
\name{thl}
\alias{thl}
\title{Heat load}
\usage{
thl(L, A, S, unit = "deg", fold = 180)
}
\arguments{
\item{L}{float value for latitude of site}

\item{A}{vector, dataframe, or raster of aspect values}

\item{S}{vector, dataframe or raster of slope values}

\item{unit}{units that aspect are in, either degrees or radians}

\item{fold}{angle along which aspect should be folded}
}
\value{
total heat load value
}
\description{
calculates heat load from input slope and aspect values
}
\examples{
\dontrun{
l <- 38.58
a <- 0:360
s <- 12.6
heatload <- thl(l, a, s, 'deg', 180)
plot(A,thl(L,A,S,fold=225))
points(A,thl(L,A,S,fold=225))
}
}
