% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatload.R
\name{thl}
\alias{thl}
\title{Heat load}
\usage{
thl(L, A, S, unit = c("deg", "rad"), fold = 180, equation = c(1, 2, 3))
}
\arguments{
\item{L}{Float value for latitude of site}

\item{A}{Vector, dataframe, or raster of aspect values in degrees or radians}

\item{S}{Vector, dataframe or raster of slope values in degrees or radians}

\item{unit}{Character vector. Units that aspect are in, either degrees or radians}

\item{fold}{Numeric. Angle along which aspect should be folded}

\item{equation}{Integer vector. The equation number from McCune and Keon 2002. Default is 3.
There three equations have slightly different uses. \emph{Eq.1} is broadest in application,
covering slopes <=90º at latitudes 0-60º N, but has the lowest R^2 when compared with
instrumental radiation observations. \emph{Eq. 2.} has higher agreement but excludes slopes >60º.
\emph{Eq. 3} holds between latitudes of ±30-60º and slope angles <= 60º and has highest R^2.}
}
\value{
total heat load value
}
\description{
Calculates heat load from input slope and aspect values
}
\note{
Heat load is a unitless index of heat load, rather than a direct estimate of
incident radiation. Their 3rd equation is valid for the narrowest input ranges
but has the highest R^2 and is suitable for many settings. The estimates are somewhat
rudimentary and do not take into account cloud cover, regional differences in the atmospheric
coefficient, or shading by adjacent topography.
}
\examples{
\dontrun{
l <- 38.58
a <- 0:360
s <- 12.6
heatload <- thl(l, a, s, 'deg', 180)
plot(A,thl(L,A,S,fold=225))
points(A,thl(L,A,S,fold=225))
}
}
\references{
McCune, B. and D. Keon (2002) Equations for potential annual direct incident
radiation and heat load. Journal of Vegetation Science 13:603-606.
}
